templateKey and Bucket added
Optional value deploymentS3Keytemplate will decide if pre-defined
template should be used or not. If provided we assume that the external
template is uploaded. The template key is to be provided as input to
maven deploy phase or pre-defined in pom.xml. The Bucket variabel is
also to be provided from the pom.xml, it is now a valid maven input
argument.